move parsing methods related to Genomics Core or processing workflows from 'illumina' module into new modules for 'gencore' and 'processing', respectively; update test module
