+++ bripipetools/parsing/__init__.py
@@ -8,6 +8,8 @@ Core (via GenLIMS).
 
 Depends on the ``util`` module.
 """
-from .illumina import (get_project_label, parse_project_label,
-                       get_library_id, get_flowcell_id, parse_flowcell_run_id,
-                       parse_fastq_filename, parse_batch_name)
+from .gencore import (get_project_label, parse_project_label,
+                      get_library_id)
+from .illumina import (get_flowcell_id, parse_flowcell_run_id,
+                       parse_fastq_filename)
+from .processing import (parse_batch_name)
